+++ src/Features/Sitecore.Pathfinder.Unicorn/UnicornExtension.cs
@@ -1,4 +1,4 @@
-﻿// © 2015 Sitecore Corporation A/S. All rights reserved.
+﻿// © 2015-2016 Sitecore Corporation A/S. All rights reserved.
 
 using Sitecore.Pathfinder.Building;
 using Sitecore.Pathfinder.Extensibility;
@@ -7,6 +7,15 @@ namespace Sitecore.Pathfinder.Unicorn
 {
     public class UnicornExtension : ExtensionBase
     {
+        public override void RemoveWebsiteFiles(IBuildContext context)
+        {
+            RemoveWebsiteAssembly(context, "Sitecore.Pathfinder.Unicorn.dll");
+            RemoveWebsiteAssembly(context, "Unicorn.dll");
+            RemoveWebsiteAssembly(context, "Rainbow.Storage.Sc.dll");
+            RemoveWebsiteAssembly(context, "Rainbow.Storage.Yaml.dll");
+            RemoveWebsiteAssembly(context, "Rainbow.dll");
+        }
+
         public override bool UpdateWebsiteFiles(IBuildContext context)
         {
             var updated = false;