+++ src/Sitecore.Pathfinder.Core/Extensibility/ExtensionBase.cs
@@ -1,4 +1,4 @@
-﻿// © 2015 Sitecore Corporation A/S. All rights reserved.
+﻿// © 2015-2016 Sitecore Corporation A/S. All rights reserved.
 
 using System.IO;
 using Sitecore.Pathfinder.Building;
@@ -9,6 +9,8 @@ namespace Sitecore.Pathfinder.Extensibility
 {
     public abstract class ExtensionBase : IExtension
     {
+        public abstract void RemoveWebsiteFiles(IBuildContext context);
+
         public abstract bool UpdateWebsiteFiles(IBuildContext context);
 
         protected virtual bool CopyProjectFileToWebsiteBinDirectory([NotNull] IBuildContext context, [NotNull] string fileName)
@@ -31,6 +33,26 @@ namespace Sitecore.Pathfinder.Extensibility
             var targetFileName = PathHelper.Combine(PathHelper.Combine(websiteDirectory, "bin"), Path.GetFileName(fileName));
 
             return context.FileSystem.CopyIfNewer(sourceFileName, targetFileName);
+        }
+
+        protected virtual void RemoveWebsiteAssembly([NotNull] IBuildContext context, [NotNull] string assemblyFileName)
+        {
+            var fileName = Path.Combine(context.WebsiteDirectory, "bin\\" + assemblyFileName);
+
+            if (!context.FileSystem.FileExists(fileName))
+            {
+                return;
+            }
+
+            try
+            {
+                context.FileSystem.DeleteFile(fileName);
+                context.Trace.TraceInformation("Removed: " + PathHelper.UnmapPath(context.WebsiteDirectory, fileName));
+            }
+            catch
+            {
+                context.Trace.TraceInformation("Failed to remove: " + PathHelper.UnmapPath(context.WebsiteDirectory, fileName));
+            }
         }
     }
 }