+++ src/Sitecore.Pathfinder.Core/Extensibility/IExtension.cs
@@ -1,4 +1,4 @@
-﻿// © 2015 Sitecore Corporation A/S. All rights reserved.
+﻿// © 2015-2016 Sitecore Corporation A/S. All rights reserved.
 
 using System.ComponentModel.Composition;
 using Sitecore.Pathfinder.Building;
@@ -9,6 +9,8 @@ namespace Sitecore.Pathfinder.Extensibility
     [InheritedExport]
     public interface IExtension
     {
+        void RemoveWebsiteFiles([NotNull] IBuildContext context);
+
         bool UpdateWebsiteFiles([NotNull] IBuildContext context);
     }
 }